requirements.txt
Flask==2.2.2
Pillow==9.2.0
img2pdf==0.4.0
pdf2image==1.16.0
gunicorn==20.1.0
